refactor config code into a sub-directory
